Move code out of mod.rs
Document the methods of `Tester`.
